Remove newDotHome beta
