+++ src/components/FloatingGPSButton/index.native.tsx
@@ -5,7 +5,6 @@ import {PressableWithoutFeedback} from '@components/Pressable';
 import {useMemoizedLazyExpensifyIcons} from '@hooks/useLazyAsset';
 import useLocalize from '@hooks/useLocalize';
 import useOnyx from '@hooks/useOnyx';
-import usePermissions from '@hooks/usePermissions';
 import useTheme from '@hooks/useTheme';
 import useThemeStyles from '@hooks/useThemeStyles';
 import Navigation from '@libs/Navigation/Navigation';
@@ -23,9 +22,6 @@ function FloatingGpsButton() {
     const {textMutedReversed} = useTheme();
     const styles = useThemeStyles();
 
-    const {isBetaEnabled} = usePermissions();
-    const isNewDotHomeEnabled = isBetaEnabled(CONST.BETAS.NEW_DOT_HOME);
-
     if (!gpsDraftDetails?.isTracking) {
         return null;
     }
@@ -37,7 +33,7 @@ function FloatingGpsButton() {
 
     return (
         <PressableWithoutFeedback
-            style={[styles.navigationTabBarFABItem, styles.ph0, styles.userSelectNone, isNewDotHomeEnabled ? styles.floatingGpsButtonAboveFab : styles.floatingGpsButton]}
+            style={[styles.navigationTabBarFABItem, styles.ph0, styles.userSelectNone, styles.floatingGpsButton]}
             accessibilityLabel={translate('gps.fabGpsTripExplained')}
             onPress={navigateToGpsScreen}
             role={CONST.ROLE.BUTTON}
