+++ src/components/Search/SearchRouter/SearchRouterPage/index.tsx
@@ -5,8 +5,7 @@ import ROUTES from '@src/ROUTES';
  * On native devices SearchRouter is served from SearchRouterPage, on web from SearchRouterModal.
  */
 function SearchRouterPage() {
-    // @TODO: Navigate to HOME when removing the newDotHome beta
-    return Navigation.navigate(ROUTES.INBOX);
+    return Navigation.navigate(ROUTES.HOME);
 }
 
 export default SearchRouterPage;
