+++ src/libs/Navigation/AppNavigator/PublicScreens.tsx
@@ -25,9 +25,9 @@ function PublicScreens() {
     const StyleUtils = useStyleUtils();
     return (
         <RootStack.Navigator screenOptions={defaultScreenOptions}>
-            {/* The structure for the HOME route has to be the same in public and auth screens. That's why the name for SignInPage is REPORTS_SPLIT_NAVIGATOR. */}
+            {/* The structure for the HOME route has to be the same in public and auth screens. That's why the name for SignInPage is SCREENS.HOME. */}
             <RootStack.Screen
-                name={NAVIGATORS.REPORTS_SPLIT_NAVIGATOR}
+                name={SCREENS.HOME}
                 options={{
                     ...defaultScreenOptions,
                     // If you want to change this, make sure there aren't any animation bugs when signing out.
