+++ src/styles/index.ts
@@ -1666,13 +1666,6 @@ const staticStyles = (theme: ThemeColors) =>
 
         floatingCameraButton: {
             position: 'absolute',
-            top: -variables.componentSizeLarge - 16,
-            right: 16,
-            zIndex: 10,
-        },
-
-        floatingCameraButtonAboveFab: {
-            position: 'absolute',
             // floatingActionButton top property value (componentSizeLarge + 16) +
             // + floatingCameraButton height (componentSizeLarge) + gap (12) = 2 * componentSizeLarge + 28
             top: 2 * -variables.componentSizeLarge - 28,
@@ -1688,15 +1681,6 @@ const staticStyles = (theme: ThemeColors) =>
         },
 
         floatingGpsButton: {
-            position: 'absolute',
-            // floatingCameraButton top property value (componentSizeLarge + 16) +
-            // + floatingGpsButton height (componentSizeLarge) + gap (12) = 2 * componentSizeLarge + 28
-            top: 2 * -variables.componentSizeLarge - 28,
-            right: 16,
-            zIndex: 10,
-        },
-
-        floatingGpsButtonAboveFab: {
             position: 'absolute',
             // floatingActionButton top property value (componentSizeLarge + 16) +
             // + floatingCameraButton height (componentSizeLarge) + gap (12) + floatingGpsButton height (componentSizeLarge) + gap (12) = 3 * variables.componentSizeLarge + 40
